+++ src/main/java/com/pekar/callofcompanions/controllers/FollowPlayerController.java
@@ -4,6 +4,7 @@ import com.mojang.logging.LogUtils;
 import com.pekar.callofcompanions.scheduler.CompanionEntryScheduler;
 import com.pekar.callofcompanions.scheduler.CompanionEntryTask;
 import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.Component;
 import org.slf4j.Logger;
 
 class FollowPlayerController extends LoadedAnimalSummonController
@@ -25,9 +26,10 @@ class FollowPlayerController extends LoadedAnimalSummonController
                 companionEntry,
                 player,
                 (ticks, entry) -> {
-                    if (ticks % 20 == 0)
+                    if (ticks % 5 == 0)
                     {
-                        setGoal(animal, player);
+                        if (animal.getNavigation().isDone())
+                            setGoal(animal, player);
                         return animal.distanceToSqr(player) < 9;
                     }
                     return false;
@@ -36,10 +38,19 @@ class FollowPlayerController extends LoadedAnimalSummonController
                     LOGGER.debug("Follow-player task completed: companionType={}, companionId={}", entry.type(), entry.uuid());
                     if (animal.distanceToSqr(player) > 10 * 10)
                     {
-                        tryTeleportAnimalTo(level, animal.getUUID(), teleportPos);
-                        setGoal(animal, player);
-                        playTeleportSound(level, animal);
-                        showAnimalTeleportParticles(level, animal);
+                        var teleported = tryTeleportAnimalTo(level, animal.getUUID(), teleportPos);
+                        if (teleported)
+                        {
+                            setGoal(animal, player);
+                            playTeleportSound(level, animal);
+                            showAnimalTeleportParticles(level, animal);
+                        }
+                        else
+                        {
+                            var name = CallCrystalHelper.buildAnimalName(entry.type(), entry.name());
+                            player.sendSystemMessage(Component.translatable("message.callofcompanions.cant_teleport", name), true);
+                            LOGGER.debug("Far teleport failed: companion couldn't find a safe place to teleport, companionType={}, companionId={}", entry.type(), entry.uuid());
+                        }
                     }
                     CallCrystalHelper.updateCompanionPos(level, companionData, entry);
                 },